Adding friendly_id for pretty URLs
Adding friendly_ids for nodes, datacenters, statuses, and roles.

Also went through and added a bunch of very unDRY before_validation methods
to most of the models. Capital letters don't appear to play nicely with the
friendly_id gem and I figured it was just easier to sanitize them on the way
in rather than clean it up later.

There's probably a better way of doing this in rails, I just don't know what
that is yet.